ponderada3/image_processor/requirements.txt
fastapi
uvicorn
pillow==10.3.0
rembg